pyproject.toml
[project]
name = "distiq-code"
version = "0.1.1"
description = "CLI proxy for optimizing AI coding assistant token usage via smart routing and caching"
authors = [
    {name = "Distiq Team", email = "hello@distiq.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ai", "coding", "cursor", "claude", "optimization", "tokens"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # CLI Framework
    "typer>=0.12.0",
    "rich>=13.7.0",

    # FastAPI Server (Proxy)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx[http2]>=0.27.0",

    # Data & Validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",

    # Utilities
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "Pillow>=10.0.0",
]
compression = [
    "llmlingua>=0.2.0",
]
all = [
    "distiq-code[ml,compression]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[project.scripts]
distiq-code = "distiq_code.cli:app"

[project.urls]
Homepage = "https://distiq.ru"
Documentation = "https://docs.distiq.ru/code"
Repository = "https://github.com/distiq/distiq-code"
Issues = "https://github.com/distiq/distiq-code/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["distiq_code"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=distiq_code --cov-report=term-missing"
